Percentageoption done
